Display hotkeys information in quizTest.js
